Merge remote-tracking branch 'origin/master' into
new_lipid_annotation_module

Conflicts:
	src/main/java/io/github/mzmine/datamodel/features/FeatureListRow.java
	src/main/java/io/github/mzmine/datamodel/features/ModularFeatureListRow.java
	src/main/java/io/github/mzmine/modules/dataprocessing/featdet_chromatogramdeconvolution/FeatureResolverTask.java
	src/main/java/io/github/mzmine/modules/dataprocessing/featdet_chromatogramdeconvolution/ResolvingUtil.java
	src/main/java/io/github/mzmine/modules/dataprocessing/featdet_imagebuilder/IImage.java
	src/main/java/io/github/mzmine/modules/dataprocessing/featdet_imagebuilder/ImageBuilderTask.java
	src/main/java/io/github/mzmine/modules/dataprocessing/filter_rowsfilter/RowsFilterTask.java
	src/main/java/io/github/mzmine/modules/dataprocessing/id_lipididentification/LipidSearchModule.java
	src/main/java/io/github/mzmine/modules/dataprocessing/id_lipididentification/lipids/lipidmodifications/LipidModificationChoiceComponent.java
	src/main/java/io/github/mzmine/modules/dataprocessing/id_lipididentification/lipidutils/LipidIdentity.java
	src/main/java/io/github/mzmine/modules/visualization/featurelisttable_modular/FeatureTableContextMenu.java
	src/main/java/io/github/mzmine/modules/visualization/spectra/simplespectra/MultiSpectraVisualizerPane.java
	src/main/java/io/github/mzmine/util/FeatureConvertors.java
	src/main/java/io/github/mzmine/util/spectraldb/parser/NistMspParser.java

+++ src/main/java/io/github/mzmine/modules/tools/msmsscore/MSMSScore.java
@@ -29,9 +29,9 @@ import java.util.Map;
 public class MSMSScore {
 
   private double score;
-  private Map<Integer, String> annotation;
+  private Map<Double, String> annotation;
 
-  public MSMSScore(double score, Map<Integer, String> annotation) {
+  public MSMSScore(double score, Map<Double, String> annotation) {
     this.score = score;
     this.annotation = annotation;
   }
@@ -40,7 +40,7 @@ public class MSMSScore {
     return score;
   }
 
-  public Map<Integer, String> getAnnotation() {
+  public Map<Double, String> getAnnotation() {
     return annotation;
   }
